refactor: synchronization server
