+++ pkg/storage/synchronization_server/interface.go
@@ -1,4 +1,4 @@
-package storage
+package synchronization_server
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 	"github.com/werf/werf/pkg/image"
 )
 
-type StagesStorageCache interface {
+type StagesStorageCacheInterface interface {
 	GetAllStages(ctx context.Context, projectName string) (bool, []image.StageID, error)
 	DeleteAllStages(ctx context.Context, projectName string) error
 	GetStagesByDigest(ctx context.Context, projectName, digest string) (bool, []image.StageID, error)
